pyproject.toml
[build-system]
requires = ["setuptools >= 64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ingest_app"
version = "0.0.1"
authors = [
  { name="Rajakumaran Arivumani", email="author@example.com" },
]
description = "Data ingestion app package"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
#license = "MIT"
#license-files = ["LICEN[CS]E*"]
dependencies = [
  # app specific 
  "pyspark == 3.5.4",
  "python-dotenv", 
  # my packages
  "utils", 
  "metadata", 
  "app_calendar", 
  "dq_app", 
  "dqml_app", 
  "dp_app", 
  "dr_app", 
  "dl_app", 
  # my packages
  #"utils @ git+https://github.com/dexplorer/utils",
  #"metadata @ git+https://github.com/dexplorer/df-metadata",
  #"app_calendar @ git+https://github.com/dexplorer/df-app-calendar",
  #"dq_app @ git+https://github.com/dexplorer/df-data-quality",
  #"dqml_app @ git+https://github.com/dexplorer/df-data-quality-ml",
  #"dp_app @ git+https://github.com/dexplorer/df-data-profile",
  #"dr_app @ git+https://github.com/dexplorer/df-data-recon",
  #"dl_app @ git+https://github.com/dexplorer/df-data-lineage",
]
keywords = ["data", "ingestion", "etl", "elt", "framework", "load"]

[project.optional-dependencies]
api = [
  "fastapi" , 
  "uvicorn[standard]", 
  ]
cli = [
  "click",
]
test = [
  # standard dev tools
  "pytest",
  "pytest-cov",
  "pylint",
  "black",
  "ipython",
]

[project.scripts]
ingest-app-cli = "ingest_app.ingest_app_cli:main"
ingest-app-api = "ingest_app.ingest_app_api:main"

[project.urls]
Homepage = "https://profile.rajakumaran.in"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/dexplorer/df-data-distribution.git"
Issues = "https://github.com/dexplorer/df-data-distribution/issues"
Changelog = "https://github.com/dexplorer/df-data-distribution/blob/master/CHANGELOG.md"
